move battle goal related code into own directory
