+++ src/lang/react.ts
@@ -1,0 +1,8 @@
+export interface NoState {
+}
+
+export interface NoProps {
+}
+
+export function noop() {
+}
